refactor: move about info to provider folder
