+++ lib/view/page_app_about.dart
@@ -21,8 +21,8 @@ import 'package:provider/provider.dart';
 import '../assets/assets.dart';
 import '../component/widget.dart';
 import '../logging/helper.dart';
-import '../model/about_info.dart';
 import '../model/contributor.dart';
+import '../provider/about_info.dart';
 import 'common/_dialog.dart';
 import 'common/contributor_tile.dart';
 import 'for_app_about/_widget.dart';
